requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
--extra-index-url https://download.pytorch.org/whl/cu118

numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
pillow==10.2.0
opencv-python==4.9.0.80
imagehash==4.3.1
albumentations==1.3.1
tqdm==4.66.1
pyyaml==6.0.1
timm==0.9.16
einops==0.7.0
grad-cam==1.5.0
# shap==0.44.0 (Removed due to MSVC requirement on Windows Python 3.12)
umap-learn==0.5.5
plotly==5.18.0
torchmetrics==1.3.0
rich==13.7.0
wandb==0.16.3
